refactored ChainReader: base.ChainReader --> chain.ChainReader
- new coordinates.chain module
- split old test_coordinates.py into test_amber_inpcrd.py and
  test_chainreader.py
- updated docs throughout (and changed/updated/fixed docs
  related to readers/writers/Trajectory API)
- added sphinx docs for chain.ChainReader
- updated CHANGELOG
