+++ package/doc/sphinx/source/documentation_pages/coordinates/chain.rst
@@ -1,0 +1,2 @@
+.. automodule:: MDAnalysis.coordinates.chain
+
