+++ package/doc/sphinx/source/documentation_pages/coordinates_modules.rst
@@ -52,5 +52,6 @@ functionality should first read the :ref:`Trajectory API`.
 
    coordinates/base
    coordinates/core
+   coordinates/chain
    coordinates/pdbextensions
